N318-011 Counterexamples for arrays.
Adding model labels for variables of array types.

Parsing of arrays was implemented in why3.

To display the content of arrays where indices are of abstract types (that
need to be projected by projection P to concrete types to by displayed),
why3 for each such array A generates new array NA with indices of
projected type and axiom stating that for all values of range type I, it
holds: P (A [I]) = NA [I].
Note that these axioms make the problem harder to prove.
The same problem is there of range of arrays. Emitting of projection
axioms for range of arrays is not yet implemented.

* gnat2why/why/gnat2why-decls.adb
(Translate_Variable): Adding counterexample labels for variables of array
type.

* gnat2why/why/gnat2why-util.adb
(Add_Counterexample_Labels): For array types, "model_trace:var_name"
and "model_projected" labels are added.