O902-004 N318-011 O619-018 Displaying return values in counterexamples.
Previously, Why3 treated all variables that were not argument of a
function inside postcondition of the function as variables corresponding
to the return value of the function. This was not correct and caused
showing values of variables internal to translation from SPARK to the
input of the prover in counterexamples in some cases.

Now, when a variable for holding the return value of a function is
generated in gnat2why, label of the form "model_trace:func_name@result"
is added to this variable and why3 uses this label to identify the
variable corresponding to the return value of a function.

Next, new label "model_vc_post", which identifies postcondition was
added. Gnatwhy3 uses the information whether the construct that triggers
VC is postcondition or not to determine whether the counterexample element
that corresponds to argument of a function has the same value as at the
time the function was called.

* gnat2why/why/gnat2why.ads
* common/vc_kinds.ads
Removed label Model_Func_Label, added label Model_VC_Post_Label
identifying postcondition.

* gnat2why/why/gnat2why-subprogram.adb
(Compute_Contract_Cases_Entry_Checks)
(Compute_Contract_Cases_Exit_Checks)
(Generate_VCs_For_Package_Elaboration)
(Generate_VCs_For_Subprogram)
(Generate_VCs_For_Task): Remove the argument for the name of the
subprogram from call to New_VC_Expr

(Generate_VCs_For_LSP)
(Generate_VCs_For_Subprogram): Generate counterexample labels for variable
holding result of a function.

* gnat2why/why/gnat2why-util.adb
(Add_Counterexample_Labels): Add information whether labels are generated
for a variable holding result of a function.

* gnat2why/why/why-atree-sprint.adb
(Print_Label): Emit current location if the label is either model_vc or
model_vc_post. That is, emit location for constructs that trigger VC.

* gnat2why/why/why-gen-expr.{adb|ads}
(New_VC_Labels): Removed argument Subprogram. If VC corresponds to
postcondition, label "model_vc_post" is emited, otherwise label
"model_vc" is emited.

* testsuite/gnatprove/tests/riposte__logic/test.out
Part of output related to counterexamples updated - there are no
superfluous counterexample elements displayed.

* why3
Changed the way how variables that should be displayed in counterexample
and are inside the term that triggers VC are collected.