O909-063 Assume dynamic invariant after refined post
We used to add the dynamic invariant of a subprogram's
effects to its refined post so that it was known when
proving the actual postcondition. It meant that this
invariant had to be reproved as part of proving the
refined post.
Now, we rather assume this dynamic invariant after the
refined postcondition.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_VCs_For_Subprogram): Dynamic invariant of
effects does not need to be checked as part of the
refined post.