O916-004 test that mimics failure on non-linux
Test to reproduce a crash when analyzing the non-linux variant of
Ada.Synchronous_Barriers library package (i.e. a variant that uses protected
objects and not native POSIX barriers; it is portable but probably less
efficient.)

Since now the crash occurs on a user-level code it clearly indicated a bug.

* testsuite/gnatprove/tests/O916-004__flow_synbar/
Test added.