O915-017 Fix missing checks in flow analysis of inlined subprograms
Because of the way inlining of subprograms works in GNATprove, it is not
equivalent to perform flow analysis of the original subprogram or the
inlined versions. In particular, some possibly failing checks on reads of
uninitialized data may be missing if only one is performed. Thus, the
previous choice to analyze only inlined version is not correct for flow
analysis, and is fixed here.

* gnat2why/flow/flow.adb, gnat2why/spark/spark_definition.adb
Call SPARK_Util.Analysis_Requested with argument With_Inlined set to True
to enforce analysis of original subprograms even when inlined for proof.

* gnat2why/spark/spark_util.adb, gnat2why/spark/spark_util.ads
(Analysis_Requested): Add parameter With_Inlined to force returning True
  on inlined subprograms.

* gnat2why/why/gnat2why-driver.adb
Call SPARK_Util.Analysis_Requested with argument With_Inlined set to False
to prevent analysis of original subprograms when inlined for proof.

* O915-017__inlined
Add test.