O910-008 Use flow analysis to get entities' global inputs
* gnat2why/flow/flow_utility.adb
(Get_Globals): Remove Globals_For_Proof parameter as is is no
longer used.
(Get_Proof_Globals): Add Keep_Constants switch to keep objects
that are constants in Why in the set of Globals.

* gnat2why/why/gnat2why-expr.adb
(Assume_Declaration_Of_Entity): New function to factor code used
for assuming the declarations of objects. It assumes the object's
dynamic invariant and as well as its value if it is a constant.
(Translate_Declaration): For nested packages declarations,
assume the dynamic invariant of initialized variables and the
package's initial condition.

* gnat2why/why/gnat2why-subprograms.adb
(Compute_Dynamic_Property_For_Inputs): Use information computed
by flow analysis to get the set of variables and constants with
variable inputs used in a subprogram, task, or package.
(Assume_Value_Of_Constants): Goes through a Why node to assume
the values and dynamic properties of constants without
variable inputs declared in it. May be possible to remove once
floating point numbers are better supported.