O922-013 wrong tag for classwide variables
When creating a classwide variable, we were wrongly setting
its tag to the tag of the classwide's specific tag.
We are now properly inheriting the tag of the variable's
initial value.

* gnat2why/why/gnat2why-expr.adb
(Assignment_Of_Object_Decl): Do not use the binder's type
to check if the assignment type is classwide, since binders
type are always specific.