O922-024 'result in functions with dispatching result
Wrong Why3 code was generated for declarations of
functions with a dispatching result and an inherited
classwide post. Indeed, when translating the post'class,
the identifier for the result had the parent's type.
We now store a proper identifier for a function's
result before translating its postcondition.

* gnat2why/why/gnat2why-expr.adb
(Transform_Attr): Used the stored identifier for 'result
attribute, even in logic.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_Subprogram_Completion)
(Translate_Expression_Function_Body): Store an adequate
identifier for the function's result.