O715-033 Fix use of default level when underlying switch is set
The previous behavior of setting the default value of level to 0 was not
satisfying when underlying switches are set. For example, when --timeout
was set, this meant --level=0 was still setting --steps=100 even when the
timeout was large, which is unexpected. Now, --level=0 is used as default
only when no underlying switch is set.

* docs/ug/appendix.rst
Update usage message.

* docs/ug/gnatprove.rst
Update doc of --level.

* docs/ug/gnatprove_by_example/Makefile.examples
Set explicitly provers cvc4,altergo,z3.

* docs/ug/gnatprove_by_example/results/perm-lemma_subprograms.prove
* docs/ug/gnatprove_by_example/results/update_list_zero.prove
Restore good results.

* gnatprove/configuration.adb
Update usage message.

* gnatprove/gnatprove.adb
(Compute_Why3_Args): Implement new behavior of default --level when
  underlying switch set. Also fix bug with invalid steps passed to provers,
  and simplify code.