O923-036 Clean ancestor field of tagged records
The anscestor field is how we model private parts
of a record coming from an ancestor.
Clean up its constructor and extractors to only
include in it parts on the ancestors that are really
private.

* gnat2why/why/why-gen-records.adb
(Count_Root_Fields_Not_In_E): Counts the number of
visible fields in E's root record type that are
private in E.
(Get_Ancestor_Part_From_Root): Call the ancestor
field constructor with the appropriate arguments.
(Declare_Ada_Record): Only declare extractors
from the ancestor for root fields that are
private in E.