O923-036 Use the same function for field extraction
In a record type, there is a private extension field
that stands for fields that will be declared in
children. When converting to the root, additionnal
fields are hidden in this extension field. When
converting back to a child, they are extracted
from the extension field.
Use the same extraction function for a given field
in all the hierarchy.