O402-017 Forbid dispatch on primitive of untagged private
The full view of an untagged private type may be tagged, in which case
primitives that are declared in the visible part of the package (hence
without visibility on the full view) cannot be declared with classwide
contracts, which lead to spurious checks that the default classwide
precondition of True was not stronger than the subprogram's precondition.
In SPARK, these subprograms should simply not be allowed in dispatching
calls, and LSP should not be checked on these.

* gnat2why/spark/spark_definition.adb
(Mark_Function_Call): Issue error on dispatching call on visible primitive
  of untagged private type.

* gnat2why/spark/spark_util.adb, gnat2why/spark/spark_util.ads
(In_Visible_Declarations): New query function.
(Is_Invisible_Dispatching_Operation): New query function to detect the
  special dispatching operations which should not be allowed to dispatch
  in SPARK.

* gnat2why/why/gnat2why-driver.adb
(Do_Generate_VCs): Do not generate LSP related VCs on such primitives.

* testsuite/gnatprove/tests/O402-017__tagged/no_primitive.ads
* testsuite/gnatprove/tests/O402-017__tagged/no_primitive_call.adb
* testsuite/gnatprove/tests/O402-017__tagged/no_primitive_call.ads
* testsuite/gnatprove/tests/O402-017__tagged/no_primitive_use.adb
* testsuite/gnatprove/tests/O402-017__tagged/test.out
* testsuite/gnatprove/tests/O402-017__tagged/test.py
Add test.