O817-002 flow - imported constant
Flow analysis used to incorrecly assume that all constants without
an initialization expression are deferred constants. This is not
the case for imported constants. We now properly process imported
constants and consider them to be initialized.

* flow_utility-initialization
(Is_Default_Initialized): Consider imported entities to be default
  initialized.

* flow-control_flow_graph
(Do_Object_Declaration): We now process imported constants.