O429-052 exclusivity checks for main subprogram
Conflicts for main subprogram accessing unsynchronized data are now
detected. Support includes main subprograms that are renamings, but not yet
main subprograms that are instances of generics.

* gnat2why/flow/flow_generated_globals.adb
(GG_Read): register the name of a task for the the main subprogram without
  the "standard__" prefix; also detect main subprograms that are renamings.

* gnat2why/utils/common_containers.ads
(To_Entity_Name): prevent calls for package and subprogram bodies, which are
  error-prone because their result is prefixed with "standard__".

* gnat2why/spark/spark_util.ads
(Might_Be_Main): restore stronger precondition; no need to call it for
  subprogram bodies.

* testsuite/gnatprove/tests/O917-095__flow_tasking_main_program/
Test extended.