O429-052 detect instance of generic as a main task
Detect instances of generic subprograms that have no parameters and are
either procedures or functions returning integers as executed by the
environment task.

Also, do not detect renamings of such subprograms as executed by the main
task, because GNAT would not allow that. (Yet, until O917-095 is solved we
will still get a warning, but it comes from the analysis of the original
subprogram).

* gnat2why/flow/flow_generated_globals.adb
(Detect_Main_Subprogram): detect subprograms that are instances of generic
  main-like subprograms.

* testsuite/gnatprove/tests/O917-095__flow_tasking_main_program/
Test extended and output updated.