N318-011 Keeping order of counter example elements.
Changes from the commit cb794190d237dbae8eed96bde7e629e07fb17b02 caused
that counterexample elements at given line were displayed in an
alphabetical order, not in the order in that they appear in the source
code.

This commit fixes this, counterexample elements are displayed in the order
in that they were collected by why3 (this order should corresponds to the
order in that they appear in SPARK source code).

(Build_Variables_Info): Store the order in that variables appear in the
counterexample before pretty printing.
(Create_Pretty_Line): Use the order in that variables appear in the
counterexample line before pretty printing to create pretty printed
counterexample line.