O924-001 Do not require SPARK_Mode on external axioms
If a package has external axioms but is not SPARK_Mode
on, we do not fail but rather ignore the axiomatization.
It is to accomodate generic instances more easily.

* gnat2why/spark/spark_util.adb
(Entity_In_Ext_Axioms): Only return True if SPARK_Mode
is On for the package.

* gnat2why/spark/spark_definition.adb
(Mark_Package_Entity): Do not check for SPARK_Mode of
packages with external axioms anymore.

* gnat2why/why/gnat2why-annotate.adb
(Syntax_Check_Pragma_Annotate_Gnatprove): Ignore
pragma Annotate for external axioms.