N318-011 Map counterexample information related to VC to correct location.
Originally, the line number of the construct that triggers VC was computed
by why3 and in gnat2why, the information mapped to this number was used to
produce one liner counterexample output. However, in general, the precise
line number of the construct that triggers VC is not known in why3. This
may be case, e.g., if the construct that trigger VC spans multiple lines
and  the VC was split. Thus, the line number computed in why3 sometimes
did not match with the line number known in gnat2why. In this case, one
liner output was not get.

To solve this problem, instead of mapping counterexample information
related to the construct that trigger VC to the line number, why3 maps it
to a dedicated index "vc_line". Gnat2why than remaps the information
to the correct line number.

gnatwhy3 stores the counterexample information
related to the construct that trigger VC in dedicated json fi

Since in gnatwhy3, the line of the construct that triggers VC is not
known, map the counterexample information related to this construct
to index "vc_line" instead of mapping it to line number.

Note that in gnatwhy3, the line of the construct that triggers VC is known
only if this construct does not span to multiple lines or if the VC is
not split.

* gnat2why/flow/flow_error_messages.adb
(Create_Pretty_Cntexmp): Call Create_Pretty_Cntexmp.Remap_VC_Info to
remap the counterexample information about the construct that triggers VC
to correct location.
(Create_Pretty_Cntexmp.Remap_VC_Info): Implementation of the remapping.

* testsuite/gnatprove/tests/riposte__records/test.out
Add one liner to the test output that was previously not got.