N318-011 Corrections in displaying counterexamples for arrays.
Unconstraint arrays are represented as records with fields elts and rt.
The former field holds the data of an array, the latter field holds the
bound of an array. This commit Disables displaying the name of the record
field elts and disables displaying the record field rt in counterexample.
That is, the content of the array is displayed in the counterexample (just
the name of the record field where it is stored is hidden), the bounds
of the array are not displayed in the counterexample.

* share/spark/theories/ada__model.mlw
Labels "model_trace:" added to not display the name of the fields elts
of the record representing unconstraint array in counterexample, labels
"model_trace:@hide_field" added to not display fields rt of this record
in counterexample.