O901-021 N318-011 Improve counterexample messages.
Update of the output of one-liner counterexamples. Now, one-liner the
following are displayed:
postcondition might fail (e.g. when Saturate'Result = 2048 and A(3) = 200)
or
postcondition might fail, cannot prove Saturate'Result < 256 (e.g. when
Saturate'Result = 2048 and A(3) = 200)

In addition, if there is a problem in getting a counterexample, the
counterexample is not displayed instead of displaying the following error
messages:
"error: cannot get location of the check"
"no information for location of the check"

* gnat2why/flow/flow_error_messages.adb
(Error_Msg_Proof.Get_Cntexmp_One_Liner): Do not display error messages
when one liner cannot be got.
(Error_Msg_Proof): Change the output of the counterexample to desired
format.

* testsuite/gnatprove/tests/O730-002__counterex_bitwise/test.out
* testsuite/gnatprove/tests/riposte__basic/test.out
* testsuite/gnatprove/tests/riposte__logic/test.out
* testsuite/gnatprove/tests/riposte__records/test.out
Update of the test output.