OA05-012 Check that ceiling priority of protected objects
If a protected object contains a procedure on which the
aspect Attach_Handler is set, then its ceiling priority
should be in Interrupt_Priority.

* gnat2why/spark/spark_util.adb
(Get_Priority_From_Protected_Type): Get the expression
of the pragma Priority or Interrupt_Priority for a
protected type if any.
(Requires_Interrupt_Priority): Returns True if the
protected type contains a procedure with a pragma
Attach_Handler.

* gnat2why/why/gnat2why-expr.adb
(Transform_Declaration): Issue a check for the priority
of a protected object if it is set by a pragma (not a
default) an if the type requires it.