N318-011 OA05-114 Declare projection for unconstrained array type.
For unconstrained array type, an abstract type is defined. There was
missing definition of a projection of values of this abstract type to
values of corresponding concrete type. This caused values of the abstract
type not be shown in counterexample. This commit fixes it by marking
function "to_array" as projection function for unconstrained arrays.

* gnat2why/why/why-gen-arrays.adb
(Declare_Unconstrained): Mark  function "to_array" as projection function.