OA19-003 N318-011 Handling of record fields not present in counterexample.
There can be some record fields missing in a record aggregate displayed in
counterexample. This is mostly because they are not relevant for the
counterexample. If there is one such field in the aggregate, it is
displayed with value ?, if there are more than one fields in the
aggregate, they are summarized by the field others. For example:

(A => ?, B => 0, C => 0)
  - if value of the field A is not known
(C => 0, others => 0)
  - if values of fields A and B are not known

* gnat2why/flow/flow_error_messages.adb
(Error_Msg_Proof.Create_Pretty_Cntexmp.Create_Pretty_File.
Create_Pretty_Line.Build_Pretty_Line.Get_Var_Or_Field_Value):
Display "?" as a value of a field that is missing in counterexample if
there are at most one such field, do not display missing fields and
summarize them with field "others" having value "?" if there are more of
them.