OA05-076 Assume dynamic invariants after Assert_And_Cut
After an Assert_And_Cut, everything is forgotten by the
tool except what is stated in the Assert_And_Cut. We
should also know that the variables modified still
abide by their dynamic invariants.
As the set of variables modified by the previous
statements of a block are quite difficult to get, we only
assume the dynamic invariant of variables mentioned in
the assert_and_cut pragma.

* gnat2why/why/gnat2why-expr.adb
(Transform_Statement_Or_Declaration_In_List): After an
Assert_And_Cut, assume the dynamic invariant of the
variables mentioned in the cut assertion.