N318-011 Enable counterexamples on spark2014 testsuite.
Enable counterexamples on spark2014 testsuite and update test outputs for
most of the tests in the public part of the testsuite. Note that outputs
of some of the tests are not updated (they need more attention).

* testsuite/gnatprove/lib/python/test_support.py
(prove_all): Remove option "--no-counterexample"
(prove_all_no_counterexample): run test with counterexamples disabled

* testsuite/gnatprove/tests/*
Updated tests.