O806-003 Add definition of expr fun to why program fun
For axioms of expression functions to be correct we need
to guard them by the ranges of their inputs. It makes
the axiom impossible to use when the solver is not
smart enough to deduce that the arguments of a call
are in range (it is often the case for floats, where the
ranges are proved by an internal analysis on bounds).
In the postcondition of the program function, we do
not need to put these guards.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_Subprogram_Program_Fun): If the subprogram is
an expression function, assume its value in its
postcondition.