OA21-004 N318-011 Disable counterexamples on selected tests.
As discussed on OA21-004, if CVC4 resource limit is reached, no
counterexample is produced on Darwin and counterexample that makes no
sense is produced on Linux. Since CVC4 does not output the information
that the resource limit was reached, it is not possible to use this
information to not output counterexamples on Linux.

Different outputs for Darwing and Linux are causing many diffs in
testsuite so counterexamples are turned off for these tests now.

Note that this is a temporary fix, I will ask CVC4 developers to solve
the problem (that outputs are different for Linux and Darwing and no
information is given when the resource limit is reached).