O929-066 Inline calls on predicate functions
Predicate functions should never be called in
the generated why as they are not expression
functions and do not have contracts.
Instead, when we encounter a call of a prediacte
function, we should replace it by its expression.

* gnat2why/why/gnat2why-expr.adb
(Dynamic_Predicate_Expression): Computes a Why3 term for
the expression contained in a dynamic predicate.
(Compute_Dynamic_Predicate): Use
Dynamic_Predicate_Expression when the predicate is
inlined.
(Transform_Expr): On a function call, if the called
entity is a predicate function, inline its value using
Dynamic_Predicate_Expression.