OA26-009 Fix of wrong AST access in pretty printing of counterexamples.
AST entity can be asked for the presence of out modifier only if it is
formal object declaration or parameter specification. Previously, all
entities were asked for that which results in wrong AST accesses and it
also caused global variables being marked with 'Old in counterexample if
the counterexample was for a postcondition of a function and the variables
were relevant to the counterexample and not modified by the function.

* gnat2why/flow/flow_error_messages.adb
(Error_Msg_Proof.Create_Pretty_Cntexmp.Create_Pretty_File.
Create_Pretty_Line.Build_Pretty_Line.Build_Variables_Info): Do not get
information whether counterexample elements have out attribute (and
do not mark them with 'Old) if they are formal object declarations or
parameter specifications.

* testsuite/gnatprove/tests/*
Update test output.