OA26-025 OA23-036 Fix of wrong AST access.
This wrong AST access happened when there was a counterexample element
of type array of record. In this case, fields and discriminants of the
record type are attempted to iterate in the array type. This commit fixes
it by checking whether the type that's fields and discriminants are
attempted to iterate can have fields or discriminants. This is workaround,
this check will not be needed when arrays of records will be supported.

* gnat2why/flow/flow_error_messages.adb
(Error_Msg_Proof.Create_Pretty_Cntexmp.Create_Pretty_File.
Create_Pretty_Line.Build_Pretty_Line.Build_Variables_Info): Check whether
the type that's fields and discriminants will be iterated can have fields
or discriminants.