OA26-037 fix wrong access to protected variables
Wrong AST access occurred when code originally written for records was used
with protected objects. There is no need to check discriminants when
accessing components of a protected object and call to Do_Discriminant_Check
was not valid.

* gnat2why/why/why-gen-records.adb
(New_Ada_Record_Access): do not call Do_Discriminant_Check for protected
  variables.