O908-014 change handling of pragma Annotate with limit-subp
On this TN, the problem was that messages were issued when gnatprove was
run with --limit-subp, even though they were suppressed by a pragma
Annotate (which was very far from the subprogram in question). The
reason was that pragma Annotate was processed only if it was directly on
the requested subprogram - but other pragmas may have an influence on
the subprogram.

Now we change and process annotation pragmas all the time. However, we
do not issue any message about useless pragmas, because this would
result in many false positives when run with --limit-subp.

* gnat2why-annotate.adb
(Generate_Useless_Pragma_Annotate_Warnings): disable when option
--limit-subp is present
(Mark_Pragma_Annotate): remove code which disabled marking of pragma
when it did not concern current subprogram (that test was too
restrictive)