OA30-004 use target interrupt_priority instead of fixed constant
The code was referring to System.Interrupt_Priority on the system where
SPARK is compiled. Instead, we should refer to the entity of the target
environment.

* gnat2why-driver.adb:
(GNAT_To_Why) unlocking of scope stack to be able to query system
  entities
* gnat2why-expr.adb
(Transform_Declaration): replace use of constant by reference to proper
  System.Interrupt_Priority entity
* gnat2why-subprogram.adb
(Check_Ceiling_Priority_Protocol): idem