OB03-034 OA19-023 Display correct values for variables in loops.
If there is both counterexample value for a variable in the first
iteration of a loop and counterexample value for the variable after the
first iteration, the former one was displayed. This is incorrect and this
commit fixes it by displaying the value after the first iteration.

Assuming that the invariant is given at the end of the loop, to handle the
first iteration of a loop, gnat2why generates why3 code with the first
iteration of the loop unrolled followed by why loop representing following
iterations of the SPARK loop. That is, for a single assignment to a
variable in the loop there, are two assignments (and WP then creates two
variables - counterexample elements). The first assignment is for the
'Loop_Entry and the second one for value after the first iteration.
These assignments have the same SPARK location and why3 cannot
distinguish between them for the present. Consequently, it reports two
counterexample elements on the same location with the same name.
Counterexample elements for a given location are given as a list (JSON
array) ordered according to the position of the element in the why3 code).
Previously, when there were two counterexample elements with the same name
and the same location, the first one was displayed. Now, it is displayed
the last one. Hence, instead of value for loop entry, the value after the
first iteration is displayed.

Values in the first iteration should be also displayed and  marked by
'Loop_Entry. This will be implemented later in scope OA19-023.

* gnat2why/flow/flow_error_messages.adb
(Create_Pretty_Cntexmp.Create_Pretty_File.Create_Pretty_Line.
Build_Pretty_Line.Build_Variables_Info.Insert_Var_Or_Field): If there
are two counterexample elements with the same name on the same line,
display the last one.

* testsuite/gnatprove/tests/*
Update tests outputs.

* testsuite/gnatprove/tests/LB28-004__loop_variant/test.py
Increase step limit to get correct counterexample.