O709-030 Issue low severity message on string concatenations
Range checks on string concatenations are likely to be unprovable
due to missing ranges on arguments of the concatenation. Issue a low
severity message in such a case, instead of the default medium severity
for unproved proof checks.

* gnat2why/flow/flow_error_messages.adb
(Error_Msg_Proof): Set message severity to low for concatenation of
  strings.

* tests/N523-032__hello/test.out
* tests/O512-022__power_and_sum/test.out
* tests/patience/test.out
* tests/spark_book/test.out
* tests/spark_io/test.out
* tests/tokeneer/test.out
Update output.