M412-032 Allow use of --limit-line and --limit-subp for generic code
Switches --limit-line and --limit-subp which are used by commands
"Prove Line" and "Prove Subprogram" in GPS and GNATbench did not work for
code inside generics. This is now fixed by not forcing -u on the file
on which the line/subprogram appears when -U is specified.
The result is that all instances of the generic are analyzed, and results
are reported for the different instances. Another patch
will follow on spark2014.py plugin to specify -U when "Prove Line" or
"Prove Subprogram" in GPS and GNATbench are used on code inside generics.

* gnat2why/spark/spark_util.adb
(Subp_Location): Update comment to justify choice of first location
  including in the case of generic instantiations.

* gnatprove/configuration.adb
(Read_Command_Line): Switches --limit-line and --limit-subp should only
  have an effect on adding -u when -U is not specified.

* testsuite/gnatprove/tests/M412-032__prove_line/main.adb
* testsuite/gnatprove/tests/M412-032__prove_line/test.opt
* testsuite/gnatprove/tests/M412-032__prove_line/test.out
* testsuite/gnatprove/tests/M412-032__prove_line/test.py
Update test and output.