OB05-007 OA28-032 Fix of emitting some of wrong why locations.
This commit fixes the following problems of emitting why locations while
printing nodes of type Why_Node_Id:
1. A node N of type Why_Node_Id can have no SPARK AST node associated. In
this case Get_Ada_Node (N) returns Empty and First_Sloc (Empty) returns
some default location and this default location was emitted.
2. W_Try_Block has never SPARK AST node associated and why location was
(attempted to be) emitted for it.

To fix these, this commit implements the following changes: (1) if node of
type Why_Node is printed, it is checked whether it has SPARK AST node
associated. If not, location is not got for such a node. (2) For nodes of
type W_Try_Block, why locations are not emitted. (3) For nodes that need
locations to be emitted, locations are emitted even if they are the same
as locations of previous nodes.

* gnat2why/why/why-atree-sprint.adb
(Print_Node): Change location only if the node being printed has
associated SPARK AST node (that is, only if the new location is
meaningful), do not emit why locations for nodes of type W_Try_Block.
(Print_Sloc_Tag): Always emit why location.

* testsutite/gnatprove/tests/*
Outputs of tests updated.