OA29-017 fix missing range check on some object declarations
SPARK did not check that the discriminants of a subtype belong to the
subtype of the corresponding discriminant in the base type. This has
been fixed now.

* gnat2why-expr.adb
(Transform_Declaration): in the case of record subtypes and protected
subtypes, check whether the discriminant constraint lies in the range of
the type of the discriminant of the base type