OB03-030 fix reference to "self" variable in axiom
The problem of this TN is a definition of a constant in a protected
subprogram like this :

  C : constant Ty := <Expression using the protected object's fields>

usually, gnat2why adds an axiom of the form

  axiom a : c = <translation of expression>

but only when the translation does not contain derefence or any.

But for protected objects, we use a "self" object which also cannot be
mentioned in axioms. So we add that to the list of things to check in
the translated expression.

* gnat2why-expr.adb
(Transform_Identifier): add Ada_Node to self Id
* gnat2why-subprograms.adb
(Generate_VCs_For_Subprogram): idem
* why-gen-terms.adb
(Has_Dereference_Or_Any): also check for reference to "self" object