minor refactoring of map update procedure
Instead of adding a singleton set to map try to insert an empty set and then
add an element to it. The original code was first searching the map with
Find and then with Insert. Now searching is only done once.

* gnat2why/why/gnat2why-error_messages.adb
(Add_It_To_Entity): refactored to avoid double searching.