minor contract for Register_VC
If Register_VC procedure is accidentally called with empty Node or Entity
gnatprove will launch why3 and then crash while trying to parse why3 output.
Investigating the cause of such delayed crashes is difficult; it is better
to check the validity of arguments when registering a VC.

* gnat2why/why/gnat2why-error_messages.{ads,adb}
(Register_VC): precondition added.