OB16-013 marking of name of a call
previously it was unnecessary to mark the "name" of a call - the called
subprogram, because it was always an identifier that was separately
marked. With protected objects, the "name" may be a more complex
expression and needs to be marked.

* spark_definition.adb
(Mark): adjust marking of N_Selected_Component case, this is sometimes
  encountered with calls to protected objects
(Mark_Call): mark the "name" of a call