OA26-021 Do not display uninitialized counterexample elements.
Do not display counterexample values for uninitialized variables and
for uninitialized arguments of functions or procedures.

Counterexample is uninitialized and will not be displayed if it has same
location as its declaration and its declaration has no initialization or
its declaration is declaration of out argument. Otherwise, counterexample
element was created for modification of a variable (and thus is
initialized) or counterexample element is a part of construct that
triggers VC (e.g., postcondition). In the latter case, counterexample
element must be also initialized, otherwise flow error will be get.

* gnat2why/flow/flow_error_messages.adb
(Error_Msg_Proof.Create_Pretty_Cntexmp.Create_Pretty_File.
Create_Pretty_Line.Build_Variables_Info): If a counterexample element is
uninitialized, filter it out from the counterexample.