OB23-014 Always rewrite real literals into machine numbers
In some cases, the frontend does not rewrite real literals into machine
numbers, something that is fine for compilation as gigi can handle it, but
not for analysis with GNATprove. GNATprove now handles this rewritting
when not performed by the frontend.

* gnat2why/spark/spark_rewrite.adb
(Rewrite_Real_Literal): New procedure to rewrite real literals.

* testsuite/gnatprove/tests/OB23-014__float_literal/test.out
* testsuite/gnatprove/tests/OB23-014__float_literal/test.py
* testsuite/gnatprove/tests/OB23-014__float_literal/u.adb
New test.