OB17-003 volatile calls in interfering contexts
Calls to volatile functions are allowed only in non-interfering contexts and
this is now checked in SPARK filter.

This is a preliminary patch: rules for internal/external calls to protected
functions still need to be checked and volatile calls as actual parameters
are not rejected.

* gnat2why/spark/spark_definition.adb
(Mark_Call): check if volatile function is called from non-interfering
  context.

* testsuite/gnatprove/tests/O526-035__protected/test.out
Violation detected so output updated.

* testsuite/gnatprove/tests/OB17-003__volatile_call/
* testsuite/gnatprove/tests/OB17-003__volatile_actual/
Test for detecting volatile calls and objects as actual parameters.

* testsuite/gnatprove/tests/O429-046__flow_tasking_cfg/foo.adb
* testsuite/gnatprove/tests/OA30-055__potentially_blocking_default_arg/
* testsuite/gnatprove/tests/O429-048__flow_.../blocking_contract.adb
* testsuite/gnatprove/tests/O429-048__flow_potentially_blocking/test.out
* testsuite/gnatprove/tests/crazyflie/modules/body/free_fall_pack.adb
* testsuite/gnatprove/tests/crazyflie/test.out
* testsuite/gnatprove/tests/heatingsystem/heatingsystem_dfa.adb
* testsuite/gnatprove/tests/heatingsystem/test.out
SPARK violations removed and outputs updated.

* testsuite/gnatprove/tests/OB12-011__polluting_abstract_state/test.out
* testsuite/gnatprove/tests/RM_MS__contracts_with_remote_state/test.out
SPARK violations difficult (or impossible?) to avoid.