OB28-002 avoid costly calls to Get_Cluster
If two entities are equal they must belong to the same cluster of a
component graph. Rely on this to avoid costly calls to Get_Cluster.

On the test related to this TN this improves performance in debug mode
from 60 seconds to 40 seconds.

* gnat2why/flow/flow_utility.adb
(Same_Component): avoid costly calls to Get_Cluster if entities are equal.