OB17-003 test output update
Volatile calls as actuals are not yet detected; the previously detected
calls were only reported due to a special-case for entry calls. The diff in
test output reflects was caused by removal of this special case from the
front end.

* testsuite/gnatprove/tests/OB12-011__polluting_abstract_state/test.out
Test output updated to match the current state of detecting volatile calls.