OC06-002 support for renaming-as-body
gnat2why crashed on renaming-as-body construct, because it assumed that any
subprogram spec with Body_To_Inline set is inlined by the frontend. However,
Body_To_Inline is also set in specs of subprograms completed with
renaming-as-body (which is clearly documented in sinfo.ads).

Now we correctly detect this case and pretend that there is no
Body_To_Inline in subprogram spec entity.

* gnat2why/spark/spark_util.adb
(Is_Local_Subprogram_Always_Inlined): now checks for renaming-as-body.
(Has_Renaming_As_Body): added to detect subprograms with renaming-as-body.

* testsuite/gnatprove/tests/OC06-002__renaming_as_body/
Test with expected output.