OC10-031 process discriminants using list
This is yet another cleanup before actually fixing the problem of this TN.

Previously we first iterated over discriminants to count them, then created
arrays big enough to hold their representation in Why3, and finally iterated
once again to populate the arrays. Now we iterate only once and append the
Why3 representations to a list as we go, which seems more natural.

Also, now we build Why3 term starting from the priority expression and then
prepending it with discriminants starting from the last one (using reverse
keyword for list iteration). This results in more natural Why3 code, where
the order of discriminants is the same as in SPARK code.

* gnat2why/why/gnat2why-expr.adb
(Transform_Declaration): code refactored to be more natural and create more
  natural Why3 representation of protected type discriminants.