OC10-037 check for Attach_Handler on private proc
Process private declarations of protected type to check for procedures with
Attach_Handler pragma.

The check is currently generated without looking at the SPARK_Mode of
private part of the spec, but this will be fixed once the SPARK_Mode status
of multi-part entities (i.e. packages, tasks and protected types) is
recorded in marking.

We should not forget that checks should be generated *only* if SPARK_Mode is
On for *both* the protected procedure and its corresponding pragma
Attach_Handler. This should change the output for tests added in this
commit.

* gnat2why/why/gnat2why-expr.adb
(Compute_Attach_Handler_Check): check also private declarations.
(Process_Declarations): procedure for scanning declaration lists.

* testsuite/gnatprove/tests/OC10-037__private_attach_handler/
Test output updated; test for not yet respected SPARK_Mode added.